+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
     - Remote sync for projectroles app setting (#533)
     - IP address based access restriction for projects (#531)
     - ``is_delegate()`` and ``is_owner_or_delegate()`` helpers for ``Project`` model
+    - Remote sync for non-owner category members (#502)
 
 Changed
 -------